% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateAnalysisElasticNet.R
\name{multivariateAnalysisElasticNet}
\alias{multivariateAnalysisElasticNet}
\title{multivariateAnalysisElasticNet
Do a multivariate elastic net analysis}
\usage{
multivariateAnalysisElasticNet(tall_df, resp_var = NULL, lambda = NULL,
  alpha = NULL)
}
\arguments{
\item{df}{Output data from from CancerCellLines::makeRespVsGeneticDataFrame}
}
\value{
data frame
}
\description{
multivariateAnalysisElasticNet
Do a multivariate elastic net analysis
}

