% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateAnalysisGBM.R
\name{multivariateAnalysisGBM}
\alias{multivariateAnalysisGBM}
\title{multivariateAnalysisGBM
Do a multivariate boosted tree analysis}
\usage{
multivariateAnalysisGBM(tall_df, resp_var = NULL, n.trees = NULL,
  interaction.depth = NULL, shrinkage = NULL)
}
\arguments{
\item{df}{Output data from from CancerCellLines::makeRespVsGeneticDataFrame}
}
\value{
data frame
}
\description{
multivariateAnalysisGBM
Do a multivariate boosted tree analysis
}

