% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateAnalysisHeatmap.R
\name{multivariateAnalysisHeatmap}
\alias{multivariateAnalysisHeatmap}
\title{multivariateAnalysisHeatmap}
\usage{
multivariateAnalysisHeatmap(caret_res_list, n = 10)
}
\arguments{
\item{caret_res_list}{list of caret train objects}

\item{n}{number of top features to plot}
}
\value{
draws a heatmap to the current device
}
\description{
Make a heatmap of the results from a list of train objects generated by multivariateAnalysisXXX functions
}

