% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateAnalysisPrep.R
\name{multivariateAnalysisPrep}
\alias{multivariateAnalysisPrep}
\title{multivariateAnalysisPrep
Converts a tall data frame from makeTallDataFrame into a wide data frame with a single response variable ready for multivariate analysis}
\usage{
multivariateAnalysisPrep(tall_df, resp_var = NULL)
}
\arguments{
\item{resp_var}{Response variable present in the data.frame - ie value of ID field when Type field is resp}

\item{df}{Output data from from CancerCellLines::makeTallDataFrame}
}
\value{
data frame
}
\description{
multivariateAnalysisPrep
Converts a tall data frame from makeTallDataFrame into a wide data frame with a single response variable ready for multivariate analysis
}

