% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateAnalysisVarImp.R
\name{multivariateAnalysisVarImp}
\alias{multivariateAnalysisVarImp}
\title{multivariateAnalysisVarImp
Extract variable importance parameters from a caret::train object}
\usage{
multivariateAnalysisVarImp(caret_res)
}
\arguments{
\item{caret_res}{Train object as produced by multivariateAnalysisElasticNet for example}
}
\value{
data frame
}
\description{
multivariateAnalysisVarImp
Extract variable importance parameters from a caret::train object
}

