% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateAnalysisVarImpPlot.R
\name{multivariateAnalysisVarImpPlot}
\alias{multivariateAnalysisVarImpPlot}
\title{multivariateAnalysisVarImpPlot
Plot variable importance parameters from a caret::train object}
\usage{
multivariateAnalysisVarImpPlot(caret_res, n = 20)
}
\arguments{
\item{caret_res}{Train object as produced by multivariateAnalysisElasticNet for example}

\item{n}{Number of features to plot}
}
\value{
ggplot object
}
\description{
multivariateAnalysisVarImpPlot
Plot variable importance parameters from a caret::train object
}

