% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUnivariateAnalysisApp.R
\name{shinyUnivariateAnalysisApp}
\alias{shinyUnivariateAnalysisApp}
\title{Univariate Shiny App}
\usage{
shinyUnivariateAnalysisApp(con, drug_df = NULL, gsc = NULL)
}
\arguments{
\item{con}{A \code{SQLiteConnection} object to the database}

\item{drug_df}{A data frame containing the drug data}

\item{gsc}{Should either be NULL or a GSC object as created by piano::loadGSC.  A common use would be gsc=loadGSC('MSigDB.gmt')}
}
\value{
Launches an interactive Shiny application
}
\description{
Takes a user supplied data frame of drug response data and does univariate analysis vs mutation status of genes
}

