% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shinyUnivariateAnalysisServer.R
\name{shinyUnivariateAnalysisServer}
\alias{shinyUnivariateAnalysisServer}
\title{shinyUnivariateAnalysisServer}
\usage{
shinyUnivariateAnalysisServer(input, output, con, drug_df = NULL,
  gsc = NULL)
}
\arguments{
\item{input}{shiny input}

\item{output}{shiny output}

\item{con}{SQLite connection object}

\item{drug_df}{drug data frame}

\item{gsc}{piano gsc object}
}
\value{
a shiny server
}
\description{
Create a shiny server for the univariate analysis shiny app
}

