% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateAnalysis.R
\name{univariateAnalysis}
\alias{univariateAnalysis}
\title{univariateAnalysis
Do a univariate analysis}
\usage{
univariateAnalysis(rvg_df)
}
\arguments{
\item{df}{Output data from from CancerCellLines::makeRespVsGeneticDataFrame}
}
\value{
data frame
}
\description{
univariateAnalysis
Do a univariate analysis
}

