% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateAnalysisFDR.R
\name{univariateAnalysisFDR}
\alias{univariateAnalysisFDR}
\title{univariateAnalysisFDR
Do a univariate analysis}
\usage{
univariateAnalysisFDR(df, mut_count_th = 1)
}
\arguments{
\item{df}{Output data from from CancerCellLineModelling::univariateAnalysis}

\item{mut_count_th}{Minimum number of mutated samples}
}
\value{
data frame
}
\description{
univariateAnalysisFDR
Do a univariate analysis
}

