% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateAnalysisMutCounts.R
\name{univariateAnalysisMutCounts}
\alias{univariateAnalysisMutCounts}
\title{univariateAnalysisMutCounts
Do mutation counting step for univariate analysis}
\usage{
univariateAnalysisMutCounts(rvg_df)
}
\arguments{
\item{df}{Output data from from CancerCellLines::makeRespVsGeneticDataFrame}
}
\value{
data frame of mutation counts
}
\description{
univariateAnalysisMutCounts
Do mutation counting step for univariate analysis
}

