% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/univariateVolcanoPlot.R
\name{univariateVolcanoPlot}
\alias{univariateVolcanoPlot}
\title{univariateVolcanoPlot
Makes a volcano plot from the output of the univariateAnalysis function}
\usage{
univariateVolcanoPlot(ua_df, pval_th = NULL, effect_th = 2,
  use_fdr = FALSE)
}
\arguments{
\item{pval_th}{p-value threshold (unlogged)}

\item{effect_th}{effect threshold (unlogged - ie 2 = 2 fold up or down)}

\item{df}{Output data from from CancerCellLinesModelling::univariateAnalysis}
}
\value{
plot
}
\description{
univariateVolcanoPlot
Makes a volcano plot from the output of the univariateAnalysis function
}

